requirements.txt
openai==1.61.1
PyMuPDF==1.25.3
python-docx==1.1.2
faiss-cpu==1.10.0
numpy==2.2.2
streamlit==1.42.0
transformers==4.48.3
sentence-transformers==3.4.1